pkgs/experimental/swarmauri_canon_http/pyproject.toml
[project]
name = "swarmauri_canon_http"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Jacob Stewart", email = "jacob@swarmauri.com" }]
dependencies = ["httpx>=0.27.0", "pytest-asyncio>=0.23.0"]

[tool.pytest.ini_options]
markers = [
    "test: standard test",
    "unit: Unit tests",
    "i9n: Integration tests",
    "r8n: Regression tests",
    "timeout: mark test to timeout after X seconds",
    "xpass: Expected passes",
    "xfail: Expected failures",
    "perf: Performance tests that measure execution time and resource usage",
]

log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

asyncio_default_fixture_loop_scope = "function"